+++ build.sh
@@ -8,13 +8,16 @@ cd "$(dirname "$(realpath "$0")")";
 # Exit on error
 set -e
 
+# Update package datasets
+R -e 'source("data-raw/data.R")'
+
 # Document and install
 R -e 'devtools::document()'
 
-# install local
-R -e 'devtools::install()'
-
 # Build package documentation
 R -e 'pkgdown::build_site(".", preview=FALSE)'
+
+# install local
+R -e 'devtools::install()'