+++ NEWS.md
@@ -1,4 +1,10 @@
-# mbthemes (development version)
+# mb.themes (development version)
+
+### v0.0.2
+
+* Added ggplot2 scales and customizations to use Bootstrap branding
+
+* Added `brand_on()` (similar to `thematic::thematic_on()` but picks up `_brand.yml` config instead.
 
 ### v0.0.1